+++ modules/home-manager/desktop/x11/theme.nix
@@ -20,7 +20,7 @@ in {
       "*color1" = "#fc618d";
       "*color2" = "#7bd88f";
       "*color3" = "#fce566";
-      "*color4" = "#fd9353";
+      "*color4" = "#6b9ce8";
       "*color5" = "#948ae3";
       "*color6" = "#5ad4e6";
       "*color7" = "#f9f9f9";
@@ -29,7 +29,7 @@ in {
       "*color9" = "#fc618d";
       "*color10" = "#7bd88f";
       "*color11" = "#fce566";
-      "*color12" = "#fd9353";
+      "*color12" = "#6b9ce8";
       "*color13" = "#948ae3";
       "*color14" = "#5ad4e6";
       "*color15" = "#f9f9f9";