+++ modules/nixos/ai/ollama.nix
@@ -18,6 +18,11 @@ in {
   };
 
   config = mkIf cfg.enable {
+    environment.variables = {
+      OLLAMA_API_BASE = "http://127.0.0.1:11434";
+      OPENAI_API_BASE = "http://localhost:11434";
+    };
+
     environment.systemPackages = [cfg.package];
 
     services.ollama = {