+++ modules/nixos/desktop/default.nix
@@ -2,7 +2,6 @@
   imports = [
     ./awesomewm.nix
     ./xfce.nix
-    ./newm.nix
     ./lightdm.nix
   ];
 }