+++ modules/nixos/stylix/default.nix
@@ -15,27 +15,7 @@ in {
     stylix = {
       enable = true;
 
-      # Define the Monokai Pro Spectrum theme inline
-      base16Scheme = {
-        scheme = "Monokai Pro Spectrum";
-        author = "Monokai";
-        base00 = "#131313";
-        base01 = "#191919";
-        base02 = "#222222";
-        base03 = "#69676c";
-        base04 = "#8b888f";
-        base05 = "#bab6c0";
-        base06 = "#f7f1ff";
-        base07 = "#f7f1ff";
-        base08 = "#fc618d";
-        base09 = "#fd9353";
-        base0A = "#fce566";
-        base0B = "#7bd88f";
-        base0C = "#5ad4e6";
-        base0D = "#948ae3";
-        base0E = "#948ae3";
-        base0F = "#fd9353";
-      };
+      base16Scheme = "${./../../../external/base16_monokai_pro/base16-spectrum.yaml}";
 
       polarity = "dark";
 
@@ -63,9 +43,9 @@ in {
       };
       icons = {
         enable = true;
-        package = pkgs.qogir-icon-theme;
-        light = "Qogir";
-        dark = "Qogir-dark";
+        package = pkgs.honor-icon-theme;
+        light = "Honor-grey";
+        dark = "Honor-grey-dark";
       };
 
       # System-level targets
@@ -87,7 +67,7 @@ in {
     # Ensure materia-theme is available system-wide
     environment.systemPackages = with pkgs; [
       materia-theme
-      qogir-icon-theme
+      honor-icon-theme
     ];
   };
 }